WFM on strawberry and cygwin

+++ t/Child.t
@@ -30,24 +30,38 @@ ok( $proc->wait, "wait" );
 ok( $proc->is_complete, "Complete" );
 is( $proc->exit_status, 0, "Exit clean" );
 
-$proc = $CLASS->new( sub { sleep 100 } )->start;
+$proc = $CLASS->new( sub { sleep 15 } )->start;
 
 my $ret = eval { $proc->say("XXX"); 1 };
 ok( !$ret, "Died, no IPC" );
 like( $@, qr/Child was created without IPC support./, "No IPC" );
-$proc->kill(2);
+if ( $^O eq 'MSWin32' ) {
+    diag( "on win32 we must wait on this process (15 seconds)" );
+    $proc->wait;
+}
+else {
+    $proc->kill(2);
+}
 
 $proc = $CLASS->new( sub {
     my $self = shift;
     $SIG{INT} = sub { exit( 2 ) };
     $self->say( "go" );
-    sleep 100;
+    sleep 15;
+    exit 2;
 }, pipe => 1 )->start;
 
 $proc->read;
 sleep 1;
-ok( $proc->kill(2), "Send signal" );
-ok( !$proc->wait, "wait" );
+
+if ( $^O eq 'MSWin32' ) {
+    diag( "on win32 we must wait on this process (15 seconds)" );
+    $proc->wait;
+}
+else {
+    ok( $proc->kill(2), "Send signal" );
+    ok( !$proc->wait, "wait" );
+}
 ok( $proc->is_complete, "Complete" );
 is( $proc->exit_status, 2, "Exit 2" );
 ok( $proc->unix_exit > 2, "Real exit" );
@@ -70,12 +84,18 @@ my $end = time;
 
 ok( $end - $start > 2, "No autoflush" );
 
-$proc = $CLASS->new( sub {
-    my $self = shift;
-    $self->detach;
-    $self->say( $self->detached );
-}, pipe => 1 )->start;
-
-is( $proc->read(), $proc->pid . "\n", "Child detached" );
+SKIP: {
+    if ($^O eq 'MSWin32') {
+        skip "detach is not available on win32", 1;
+    }
+    else {
+        $proc = $CLASS->new( sub {
+            my $self = shift;
+            $self->detach;
+            $self->say( $self->detached );
+        }, pipe => 1 )->start;
+        is( $proc->read(), $proc->pid . "\n", "Child detached" );
+    }
+}
 
 done_testing;
